Move graphql modules to client/src/api
