+++ .gitignore
@@ -52,7 +52,7 @@ server/dist
 client/.babelrc
 
 # This file is automatically generated when running `yarn client/generate-graphql-types`
-client/graphqlOperationResultTypes.ts
+client/src/api/types.ts
 
 # This file is automatically generated when running `yarn client/generate-graphql-schema`
 graphqlSchema.json
