+++ package.json
@@ -15,7 +15,7 @@
     "client/clean": "rm -rf ./client/dist",
     "client/stats": "STATS=1 NO_PERF_CHECKS=1 NODE_ENV=production webpack --config client/webpack.config.js --json > stats.json",
     "client/graphql/schema": "apollo-codegen introspect-schema ${API_ENDPOINT} --output ./graphqlSchema.json",
-    "client/graphql/types": "apollo-codegen generate 'client/src/**/*.ts{,x}' --target typescript --output client/graphqlOperationResultTypes.ts",
+    "client/graphql/types": "apollo-codegen generate 'client/src/**/*.ts{,x}' --target typescript --output client/src/api/types.ts",
     "client/graphql/dev": "nodemon --watch client/src --ext ts,tsx --exec 'run-s client/graphql/schema client/graphql/types'",
     "client/graphql/build": "run-s client/graphql/schema client/graphql/types",
     "server/static-server": "node server/dist/staticServer.js",
